Move scripts to bin directory
Move scripts to a bin directory.  This is standard for Node.js
packages.  Unlike most Node.js packages, these scripts are in src/bin
rather than bin, but with the rearrangement on packaging, they will
end up in bin and, should this project ever get rearranged to be more
standard, the location of the bin will not change.

Existing scripts remain in place with a deprecation warning added.
