+++ src/bin/start.ts
@@ -1,11 +1,9 @@
 #!/usr/bin/env node
 
 import "source-map-support/register";
-import { automationClient } from "./automationClient";
-import { loadConfiguration } from "./configuration";
-import { enableDefaultScanning } from "./scan";
-
-console.warn("[WARN] This script is deprecated, use 'bin/start.js'");
+import { automationClient } from "../automationClient";
+import { loadConfiguration } from "../configuration";
+import { enableDefaultScanning } from "../scan";
 
 try {
     loadConfiguration()
